+++ arch/arm/src/stm32/stm32_spi.c
@@ -1797,9 +1797,9 @@ static void spi_exchange(FAR struct spi_dev_s *dev, FAR const void *txbuffer,
 
 #ifdef CONFIG_STM32_DMACAPABLE
   if ((txbuffer && priv->txbuf == 0 &&
-      !stm32_dmacapable((uint32_t)txbuffer, nwords, priv->txccr)) ||
+      !stm32_dmacapable((uintptr_t)txbuffer, nwords, priv->txccr)) ||
       (rxbuffer && priv->rxbuf == 0 &&
-       !stm32_dmacapable((uint32_t)rxbuffer, nwords, priv->rxccr)))
+       !stm32_dmacapable((uintptr_t)rxbuffer, nwords, priv->rxccr)))
     {
       /* Unsupported memory region fall back to non-DMA method. */
